+++ src/Watch/Stat.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\FileWatch\Watch;
 
-use Innmind\FileWatch\{
-    Watch,
-    Ping,
-};
+use Innmind\FileWatch\Ping;
 use Innmind\Url\Path;
 use Innmind\Server\Control\Server\{
     Processes,
@@ -15,7 +12,7 @@ use Innmind\Server\Control\Server\{
 use Innmind\TimeWarp\Halt;
 use Innmind\TimeContinuum\Period;
 
-final class Stat implements Watch
+final class Stat
 {
     private Processes $processes;
     private Halt $halt;
@@ -31,7 +28,6 @@ final class Stat implements Watch
         $this->period = $period;
     }
 
-    #[\Override]
     public function __invoke(Path $file): Ping\OutputDiff
     {
         if (\PHP_OS === 'Linux') {